Basic Idea of SQL and ER Diagram
It can be modified later
Table table is removed. It is very hard to execute to update realtime seats status.
Tax table is only for back-end programming, so it needs not to connect to other table.
Menu table is exchanged with Recipe table. We can just use menu_name rows as actual menu titles.